# decant-standalone/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Decant API
  description: |
    RESTful API for managing hierarchical knowledge graphs with AI-powered classification and enrichment.

    Decant provides dual-hierarchy organization (functional and organizational), automatic content
    classification using GPT-4o-mini, and real-time updates via Server-Sent Events.

    ## Key Features
    - Dual hierarchy organization (function & organization)
    - AI-powered content classification
    - Phase 2 enrichment for deep analysis
    - Full-text search with advanced filters
    - Real-time notifications via SSE
    - Backup and restore capabilities
  version: 1.0.0
  contact:
    name: API Support
  license:
    name: MIT
servers:
  - url: http://localhost:3000
    description: Local development server
  - url: http://localhost:{port}
    description: Configurable port
    variables:
      port:
        default: '3000'
        description: Server port
...
paths:
...
  /api/nodes/{id}/history:
    get:
      tags: [Audit]
      summary: Get node history
      description: Get code change history for a specific node
      operationId: getNodeHistory
      parameters:
        - $ref: '#/components/parameters/NodeIdParam'
        - name: hierarchyType
          in: query
          description: Filter by hierarchy type
          schema:
            type: string
            enum: [function, organization]
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NodeHistoryResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # ============================================================
  # Hierarchy API
  # ============================================================
  /api/hierarchy/tree/{view}:
    get:
      tags: [Hierarchy]
      summary: Get hierarchy tree
      description: Retrieve the complete hierarchy tree for a view
      operationId: getHierarchyTree
      parameters:
        - $ref: '#/components/parameters/HierarchyViewParam'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TreeNode'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/hierarchy/segments:
    get:
      tags: [Hierarchy]
      summary: Get all segments
      description: List all functional segments (top-level categories)
      operationId: getSegments
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Segment'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/hierarchy/organizations:
    get:
      tags: [Hierarchy]
      summary: Get all organizations
      description: List all organizations (top-level organizational entities)
      operationId: getOrganizations
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Organization'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/hierarchy/code/{view}/{code}:
    get:
      tags: [Hierarchy]
      summary: Get node by hierarchy code
      description: Retrieve a node using its hierarchy code
      operationId: getNodeByCode
      parameters:
        - $ref: '#/components/parameters/HierarchyViewParam'
        - name: code
          in: path
          required: true
          description: Hierarchy code (e.g., A.LLM.T.1)
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Node'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/hierarchy/path/{view}/{nodeId}:
    get:
      tags: [Hierarchy]
      summary: Get ancestry path
      description: Get the full ancestry path for a node
      operationId: getAncestryPath
      parameters:
        - $ref: '#/components/parameters/HierarchyViewParam'
        - $ref: '#/components/parameters/NodeIdParam'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AncestryPathResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/hierarchy/subtree/{view}/{path}:
    get:
      tags: [Hierarchy]
      summary: Get subtree
      description: Retrieve a subtree at a specific hierarchy path
      operationId: getSubtree
      parameters:
        - $ref: '#/components/parameters/HierarchyViewParam'
        - name: path
          in: path
          required: true
          description: Hierarchy path prefix (e.g., A.LLM)
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubtreeResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/hierarchy/stats/{view}:
    get:
      tags: [Hierarchy]
      summary: Get hierarchy statistics
      description: Get depth statistics and cache information
      operationId: getHierarchyStats
      parameters:
        - $ref: '#/components/parameters/HierarchyViewParam'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HierarchyStatsResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # ============================================================
  # Search API
  # ============================================================
...
components:
  parameters:
    NodeIdParam:
      name: id
      in: path
      required: true
      description: Node UUID
      schema:
        type: string
        format: uuid
    HierarchyViewParam:
      name: view
      in: path
      required: true
      description: Hierarchy view type
      schema:
        type: string
        enum: [function, organization]
...
  schemas:
...
    Node:
      type: object
      required:
        - id
        - title
        - url
        - source_domain
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        url:
          type: string
          format: uri
        source_domain:
          type: string
        company:
          type: string
          nullable: true
        phrase_description:
          type: string
          nullable: true
        short_description:
          type: string
          nullable: true
        logo_url:
          type: string
          format: uri
          nullable: true
        ai_summary:
          type: string
          nullable: true
        extracted_fields:
          type: object
          additionalProperties: true
        metadata_tags:
          type: array
          items:
            type: string
        key_concepts:
          type: array
          items:
            type: string
        function_parent_id:
          type: string
          format: uuid
          nullable: true
        organization_parent_id:
          type: string
          format: uuid
          nullable: true
        function_code:
          type: string
          nullable: true
        organization_code:
          type: string
          nullable: true
        phase2_completed:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    TreeNode:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        type:
          type: string
        code:
          type: string
        url:
          type: string
          format: uri
        logo_url:
          type: string
          format: uri
        children:
          type: array
          items:
            $ref: '#/components/schemas/TreeNode'
    Segment:
      type: object
      properties:
        code:
          type: string
        name:
          type: string
        description:
          type: string
    Organization:
      type: object
      properties:
        code:
          type: string
        name:
          type: string
        description:
          type: string

    # ============================================================
    # Response Schemas
    # ============================================================
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        code:
          type: string
        details:
          type: string
...
    NodeHistoryResponse:
      type: object
      properties:
        nodeId:
          type: string
          format: uuid
        changes:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              hierarchyType:
                type: string
              oldCode:
                type: string
                nullable: true
              newCode:
                type: string
                nullable: true
              changeType:
                type: string
              reason:
                type: string
              triggeredBy:
                type: string
              changedAt:
                type: string
                format: date-time
              relatedNodes:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                    title:
                      type: string
    AncestryPathResponse:
      type: object
      properties:
        nodeId:
          type: string
          format: uuid
        view:
          type: string
        ancestors:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              title:
                type: string
              code:
                type: string
              type:
                type: string
        depth:
          type: integer
    SubtreeResponse:
      type: object
      properties:
        path:
          type: string
        view:
          type: string
        nodes:
          type: array
          items:
            $ref: '#/components/schemas/TreeNode'
        count:
          type: integer
    HierarchyStatsResponse:
      type: object
      properties:
        view:
          type: string
        totalNodes:
          type: integer
        maxDepth:
          type: integer
        depthBreakdown:
          type: object
          additionalProperties:
            type: integer
        cache:
          type: object
          properties:
            function:
              type: integer
            organization:
              type: integer
...
  responses:
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    NotFound:
      description: Not Found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    InternalServerError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key authentication (future enhancement)